+++ src/main/java/com/example/practica6rest/service/impl/ProductServiceImpl.java
@@ -3,14 +3,17 @@ package com.example.practica6rest.service.impl;
 import com.example.practica6rest.model.Product;
 import com.example.practica6rest.repository.ProductRepository;
 import com.example.practica6rest.service.ProductService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 
 @Service
+@Slf4j
 public class ProductServiceImpl implements ProductService {
 
     private final ProductRepository productRepository;
@@ -22,18 +25,22 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public ResponseEntity<Product> registry(Product newProduct) {
+        //TODO logic
         Product product = productRepository.save(newProduct);
+        log.info(String.valueOf(product));
 
         return new ResponseEntity<>(product, HttpStatus.CREATED);
     }
 
     @Override
-    public List<Product> getProduct() {
+    public List<Product> getAll() {
+        //TODO logic
         return productRepository.findAll();
     }
 
     @Override
     public Product update(Product newProduct, Long id) {
+        //TODO logic
         return productRepository.findById(id)
                 .map(product -> {
                     product.setProduct_name(newProduct.getProduct_name());
@@ -41,14 +48,15 @@ public class ProductServiceImpl implements ProductService {
                     product.setImg_url(newProduct.getImg_url());
                     return productRepository.save(product);
                 })
-                .orElseGet(() -> {
-                    newProduct.setId(id);
-                    return productRepository.save(newProduct);
-                });
+                .orElseThrow(() -> new EntityNotFoundException("NO SUCH DESK"));
     }
 
     @Override
     public void delete(Long id) {
+        //TODO logic
+        if (!productRepository.existsById(id)) {
+            throw new EntityNotFoundException("NO SUCH PRODUCT");     //@Todo Exceptions
+        }
         productRepository.deleteById(id);
     }
 }
